removed explicit CMSSW dependence

+++ exercises/skim_test/run.sh
@@ -4,7 +4,7 @@ print_help()
 {
     echo "Usage:"
     echo ""
-    echo "  sh run.sh [--input_file=/path/to/nanoaod.root] [--package=/path/to/package.tar.gz] [--no_skim]"
+    echo "  sh run.sh [--input_file=/path/to/nanoaod.root] [--package=/path/to/package.tar.gz] [--no_skim] [--debug]"
     echo ""
     exit 0
 }
@@ -12,6 +12,8 @@ print_help()
 NANOAODPATH=""
 PACKAGE=""
 SKIM=true
+DEBUG=false
+PYTHON2=false
 
 # Parse arguments
 for arg in "$@"; do
@@ -22,6 +24,8 @@ for arg in "$@"; do
         --input_file) NANOAODPATH=$val;;
         --package) PACKAGE=$(realpath $val);;
         --no_skim) SKIM=false;;
+        --debug) DEBUG=true;;
+        --python2) PYTHON2=true;;
     esac
 done
 
@@ -43,81 +47,71 @@ else
     echo ""
 fi
 
-if [[ -f /opt/cms/cmsset_default.sh ]]; then
-    # On a CMSSW container
-    source /opt/cms/cmsset_default.sh
-elif [[ -f /cvmfs/cms.cern.ch/cmsset_default.sh ]]; then
-    # On a machine with cvmfs
-    source /cvmfs/cms.cern.ch/cmsset_default.sh
-else
-    echo "ERROR: no known paths to cmsset_default.sh exist!"
-    exit
-fi
-
-SCRAMARCH=slc7_amd64_gcc700
-CMSSWVERSION=CMSSW_10_6_25
-
 # Go to a working directory that will be cleaned afterwards
 ORIG_DIR=$PWD
 mkdir -p temp_dir
 cd temp_dir
 
-# Setup environment (If already exists it will leave them be)
-export SCRAM_ARCH=${SCRAMARCH} && scramv1 project CMSSW ${CMSSWVERSION}
-cd ${CMSSWVERSION}/src/
-
 if [[ -f $PACKAGE ]]; then
     tar xf $PACKAGE # It will overwrite if necessary
+    cd NanoAODTools
+    bash standalone/env_standalone.sh build
+    source standalone/env_standalone.sh
 else
-    # git clone nanoAOD-tools
-    git clone https://github.com/cms-nanoAOD/nanoAOD-tools.git PhysicsTools/NanoAODTools
-    cd PhysicsTools/NanoAODTools # temp_dir/CMSSW_10_2_13/src/PhysicsTools/NanoAODTools
-    git checkout e963c70
+    git clone https://github.com/jkguiang/nanoAOD-tools.git NanoAODTools
+    cd NanoAODTools
+    bash standalone/env_standalone.sh build
+    source standalone/env_standalone.sh
 
     # Copy the extra files to where they need to go
     cp ${ORIG_DIR}/extras/* python/postprocessing/examples
 
     # Get the git information
-    cd $ORIG_DIR
-    git diff > temp_dir/${CMSSWVERSION}/src/PhysicsTools/NanoAODTools/gitdiff.txt
-    git status > temp_dir/${CMSSWVERSION}/src/PhysicsTools/NanoAODTools/gitstatus.txt
-    git rev-parse HEAD > temp_dir/${CMSSWVERSION}/src/PhysicsTools/NanoAODTools/githash.txt
+    git diff > gitdiff.txt
+    git status > gitstatus.txt
+    git rev-parse HEAD > githash.txt
 
-    # Compile with CMSSW
-    cd temp_dir/${CMSSWVERSION}/src
-    cmsenv
-    cd PhysicsTools/NanoAODTools/
-    scram b -j
+    # Check if ROOT is on $PATH
+    if ! command -v root &> /dev/null; then
+        echo "ERROR: ROOT cannot be found and is needed to set up NanoCORE"
+        exit 0
+    fi
 
     # Setup NanoCORE
     git clone https://github.com/cmstas/NanoTools
-    cd NanoTools/NanoCORE # temp_dir/CMSSW_10_2_13/src/PhysicsTools/NanoAODTools/NanoTools/NanoCORE
+    cd NanoTools/NanoCORE
     make clean
     make -j
-    cd ../../../../ # temp_dir/CMSSW_10_2_13/src/
+    cd ${ORIG_DIR}/temp_dir
 
     # Tar the PhysicsTools directory
-    tar -chJf package.tar.gz PhysicsTools \
-        --exclude="PhysicsTools/NanoAODTools/.git" \
-        --exclude="PhysicsTools/NanoAODTools/data" \
-        --exclude="PhysicsTools/NanoAODTools/python/postprocessing/data" \
-        --exclude="PhysicsTools/NanoAODTools/NanoTools/.git" \
-        --exclude="PhysicsTools/NanoAODTools/package.tar.gz"
+    tar -chJf package.tar.gz NanoAODTools \
+        --exclude="NanoAODTools/.git" \
+        --exclude="NanoAODTools/data" \
+        --exclude="NanoAODTools/python/postprocessing/data" \
+        --exclude="NanoAODTools/NanoTools/.git" \
+        --exclude="NanoAODTools/package.tar.gz"
 
     mv package.tar.gz $ORIG_DIR/
+    cd NanoAODTools
 fi
 
 if [[ "$SKIM" == "true" ]]; then
-    cd PhysicsTools/NanoAODTools
-    eval `scramv1 runtime -sh`
-    scram b -j
-
-    python scripts/nano_postproc.py \
-        ./ \
-        ${NANOAODPATH} \
-        -b python/postprocessing/examples/keep_and_drop.txt \
-        -I PhysicsTools.NanoAODTools.postprocessing.examples.vbsHwwSkimModule \
-        vbsHwwSkimModuleConstr
+    if [[ "$PYTHON2" == "true" ]]; then
+        python2 scripts/nano_postproc.py \
+            ./ \
+            ${NANOAODPATH} \
+            -b python/postprocessing/examples/keep_and_drop.txt \
+            -I PhysicsTools.NanoAODTools.postprocessing.examples.vbsHwwSkimModule \
+            vbsHwwSkimModuleConstr
+    else
+        python3 scripts/nano_postproc.py \
+            ./ \
+            ${NANOAODPATH} \
+            -b python/postprocessing/examples/keep_and_drop.txt \
+            -I PhysicsTools.NanoAODTools.postprocessing.examples.vbsHwwSkimModule \
+            vbsHwwSkimModuleConstr
+    fi
 
     # Copy back the output to parent directory
     BASENAMEWITHEXT=$(basename ${NANOAODPATH})
@@ -125,6 +119,8 @@ if [[ "$SKIM" == "true" ]]; then
     mv ${BASENAME}_Skim.root $ORIG_DIR/${BASENAME}_skimmed.root
 fi
 
-# Clean up
 cd $ORIG_DIR
-rm -rf temp_dir
+if [[ "$DEBUG" == "false" ]]; then
+    # Clean up
+    rm -rf temp_dir
+fi
